+++ pkg/microservice/aslan/core/workflow/service/workflow/job/job.go
@@ -132,6 +132,8 @@ func InitJobCtl(job *commonmodels.Job, workflow *commonmodels.WorkflowV4) (JobCt
 		resp = &UpdateEnvIstioConfigJob{job: job, workflow: workflow}
 	case config.JobBlueKing:
 		resp = &BlueKingJob{job: job, workflow: workflow}
+	case config.JobApproval:
+		resp = &ApprovalJob{job: job, workflow: workflow}
 	default:
 		return resp, fmt.Errorf("job type not found %s", job.JobType)
 	}
